\name{GENEapply}
\alias{GENEapply}
\docType{package}
\title{
An internal C++ function
}
\description{
Internal function that computes the joint score test statistic over all the SNPs for all the genes
}
\author{
Chaitanya R. Acharya, Andrew S. Allen
Maintainer: Chaitanya Acharya<c.acharya@duke.edu>
}
\references{
Chaitanya R. Acharya, Kouros Owzar and Andrew S. Allen; Joint analysis of genotype and tissue-specific variability using a novel score test to map eQTL (Manuscript submitted)
}
