\name{RowSums}
\alias{RowSums}

\docType{package}
\title{
An internal C++ function
}
\description{
Internal function that computes row sums.
}

\author{
Chaitanya R. Acharya, Andrew S. Allen
Maintainer: Chaitanya Acharya<c.acharya@duke.edu>
}
\references{
Chaitanya R. Acharya, Kouros Owzar and Andrew S. Allen; Joint analysis of genotype and tissue-specific variability using a novel score test to map eQTL (Manuscript submitted)
}


