\name{snpOUT}
\alias{snpOUT}
\docType{package}
\title{
An internal C++ function
}
\description{
Internal function that computes the joint score test statistic over all the SNPs for a given gene
}
\details{
\tabular{ll}{
Package: \tab JAGUAR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-28\cr
License: \tab GPL-2\cr
}
}
\author{
Chaitanya R. Acharya, Andrew S. Allen
Maintainer: Chaitanya Acharya<c.acharya@duke.edu>
}
\references{
Chaitanya R. Acharya, Kouros Owzar and Andrew S. Allen; Joint analysis of genotype and tissue-specific variability using a novel score test to map eQTL (Manuscript submitted)
}
